Working GPU driver + programs

+++ src/minix3.c
@@ -27,7 +27,7 @@ static uint8_t *zone_bitmap;
 
 void debug_dir_entry(DirEntry entry) {
     debugf("Entry: %u `", entry.inode);
-    for (uint8_t i=0; i<60; i++) {
+    for (uint32_t i=0; i<60; i++) {
         #ifdef minix3_DEBUG
         if (entry.name[i])
             infof("%c", entry.name[i]);
@@ -205,7 +205,7 @@ void debug_inode(VirtioDevice *block_device, uint32_t i) {
     debugf("   mtime: %d\n", inode.mtime);
     debugf("   ctime: %d\n", inode.ctime);
     debugf("   zones[]:\n");
-    for (uint8_t j=0; j<10; j++) {
+    for (uint32_t j=0; j<10; j++) {
         debugf("   zones[%d] = %x (*1024 = %x)\n", j, inode.zones[j], inode.zones[j] * 1024);
     }
 }
@@ -776,7 +776,7 @@ void minix3_get_data(VirtioDevice *block_device, uint32_t inode, uint8_t *data, 
             uint32_t indirect_zones[minix3_get_zone_size(block_device) / sizeof(uint32_t)];
             minix3_get_zone(block_device, indirect_zone, (uint8_t*)indirect_zones);
 
-            for (uint8_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
+            for (uint32_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
                 uint32_t zone = indirect_zones[indirect_zone];
                 if (zone == 0) continue;
                 debugf("Reading double indirect zone %d\n", zone);
@@ -895,7 +895,7 @@ void minix3_put_data(VirtioDevice *block_device, uint32_t inode, uint8_t *data, 
 
     // Now, get the data
     // The first 7 zones are direct zones
-    for (uint8_t direct_zone=0; direct_zone<7; direct_zone++) {
+    for (uint32_t direct_zone=0; direct_zone<7; direct_zone++) {
         uint32_t zone = inode_data.zones[direct_zone];
         if (zone == 0) {
             debugf("No direct zone %d\n", zone);
@@ -953,7 +953,7 @@ void minix3_put_data(VirtioDevice *block_device, uint32_t inode, uint8_t *data, 
         uint32_t indirect_zones[minix3_get_zone_size(block_device) / sizeof(uint32_t)];
         minix3_get_zone(block_device, inode_data.zones[7], (uint8_t*)indirect_zones);
 
-        for (uint8_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
+        for (uint32_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
             uint32_t zone = indirect_zones[indirect_zone];
             if (zone == 0) continue;
             debugf("Writing indirect zone %d\n", zone);
@@ -1007,14 +1007,14 @@ void minix3_put_data(VirtioDevice *block_device, uint32_t inode, uint8_t *data, 
         // We're done
         minix3_get_zone(block_device, inode_data.zones[8], (uint8_t*)double_indirect_zones);
 
-        for (uint8_t double_indirect_zone=0; double_indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); double_indirect_zone++) {
+        for (uint32_t double_indirect_zone=0; double_indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); double_indirect_zone++) {
             uint32_t indirect_zone = double_indirect_zones[double_indirect_zone];
             if (indirect_zone == 0) continue;
 
             uint32_t indirect_zones[minix3_get_zone_size(block_device) / sizeof(uint32_t)];
             minix3_get_zone(block_device, indirect_zone, (uint8_t*)indirect_zones);
 
-            for (uint8_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
+            for (uint32_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
                 uint32_t zone = indirect_zones[indirect_zone];
                 if (zone == 0) continue;
                 debugf("Writing double indirect zone %d\n", zone);
@@ -1067,19 +1067,19 @@ void minix3_put_data(VirtioDevice *block_device, uint32_t inode, uint8_t *data, 
         uint32_t triple_indirect_zones[minix3_get_zone_size(block_device) / sizeof(uint32_t)];
         minix3_get_zone(block_device, inode_data.zones[9], (uint8_t*)triple_indirect_zones);
 
-        for (uint8_t triple_indirect_zone=0; triple_indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); triple_indirect_zone++) {
+        for (uint32_t triple_indirect_zone=0; triple_indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); triple_indirect_zone++) {
             uint32_t double_indirect_zone = triple_indirect_zones[triple_indirect_zone];
             if (double_indirect_zone == 0) continue;
             uint32_t double_indirect_zones[minix3_get_zone_size(block_device) / sizeof(uint32_t)];
             minix3_get_zone(block_device, double_indirect_zone, (uint8_t*)double_indirect_zones);
 
-            for (uint8_t double_indirect_zone=0; double_indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); double_indirect_zone++) {
+            for (uint32_t double_indirect_zone=0; double_indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); double_indirect_zone++) {
                 uint32_t indirect_zone = double_indirect_zones[double_indirect_zone];
                 if (indirect_zone == 0) continue;
                 uint32_t indirect_zones[minix3_get_zone_size(block_device) / sizeof(uint32_t)];
                 minix3_get_zone(block_device, indirect_zone, (uint8_t*)indirect_zones);
 
-                for (uint8_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
+                for (uint32_t indirect_zone=0; indirect_zone<minix3_get_zone_size(block_device) / sizeof(uint32_t); indirect_zone++) {
                     uint32_t zone = indirect_zones[indirect_zone];
                     if (zone == 0) continue;
                     debugf("Writing triple indirect zone %d\n", zone);
